further updates

+++ server/controllers/product.controller.js
@@ -26,7 +26,7 @@ export const createProduct = async (req, res) => {
 
     try{
         await newProduct.save();
-        //console.log("Product saved: ", product, typeof product);
+        console.log("Product saved:", product);
         res.status(201).json({ success: true, data: newProduct});
     } catch (error) {
         console.error(`Error in creating product ${error.message}`);
@@ -52,11 +52,16 @@ export const updateProduct = async (req, res) => {
 export const deleteProduct = async (req, res) => {
     const {id} = req.params;
     console.log("id:", id);
+
+    if (!mongoose.Types.ObjectId.isValid(id)){
+        return res.status(404).json({ success: false, message: "Invalid Product Id"});
+    }
+
     try {
         await Product.findByIdAndDelete(id);
         res.status(200).json({ success: true, message: "Product deleted"});
     }   catch (error) {
         console.log("error deleting products", error.message);
-        res.status(404).json({ success: false, message: "Product not found" });
+        res.status(500).json({ success: false, message: "Server error" });
     }
 }
